share some functionality
