+++ algorithms/active/adt/src/main/java/de/learnlib/algorithm/adt/learner/ADTAdaptiveQuery.java
@@ -19,28 +19,18 @@ import de.learnlib.algorithm.adt.adt.ADTNode;
 import de.learnlib.algorithm.adt.automaton.ADTState;
 import de.learnlib.algorithm.adt.automaton.ADTTransition;
 import de.learnlib.algorithm.adt.util.ADTUtil;
-import de.learnlib.query.AdaptiveQuery;
 import net.automatalib.word.Word;
 
-class ADTAdaptiveQuery<I, O> implements AdaptiveQuery<I, O> {
+class ADTAdaptiveQuery<I, O> extends AbstractAdaptiveQuery<I, O> {
 
-    private ADTNode<ADTState<I, O>, I, O> currentADTNode;
     private final ADTTransition<I, O> transition;
-
-    /**
-     * The index of the access sequence of the transition. If equal to the length of the access sequence the actual
-     * transition input symbol is the current symbol. If larger than the length of the access sequence, the symbol
-     * should be fetched from the ADT nodes.
-     */
-    private int asIndex;
     private final Word<I> accessSequence;
 
-    private ADTNode<ADTState<I, O>, I, O> tempADTNode;
-    private O tempOut;
+    private int asIndex;
 
     ADTAdaptiveQuery(ADTTransition<I, O> transition, ADTNode<ADTState<I, O>, I, O> root) {
+        super(root);
         this.transition = transition;
-        this.currentADTNode = root;
         this.accessSequence = transition.getSource().getAccessSequence();
         this.asIndex = 0;
     }
@@ -54,58 +44,33 @@ class ADTAdaptiveQuery<I, O> implements AdaptiveQuery<I, O> {
             }
 
             return this.accessSequence.getSymbol(this.asIndex);
-
         } else {
-            return this.currentADTNode.getSymbol();
+            return super.currentADTNode.getSymbol();
         }
     }
 
     @Override
     public Response processOutput(O out) {
-
         if (this.asIndex <= this.accessSequence.length()) {
             if (this.asIndex == this.accessSequence.length()) {
                 this.transition.setOutput(out);
             }
 
             // if the ADT only consists of a leaf, we just set the transition output
-            if (ADTUtil.isLeafNode(this.currentADTNode)) {
+            if (ADTUtil.isLeafNode(super.currentADTNode)) {
                 return Response.FINISHED;
             }
 
             asIndex++;
             return Response.SYMBOL;
         } else {
-            final ADTNode<ADTState<I, O>, I, O> succ = currentADTNode.getChild(out);
-
-            if (succ == null) {
-                this.tempOut = out;
-                this.tempADTNode = currentADTNode;
-                return Response.FINISHED;
-            } else if (ADTUtil.isResetNode(succ)) {
-                this.currentADTNode = succ.getChild(null);
-                this.asIndex = 0;
-                return Response.RESET;
-            } else if (ADTUtil.isSymbolNode(succ)) {
-                this.currentADTNode = succ;
-                return Response.SYMBOL;
-            } else {
-                this.currentADTNode = succ;
-                return Response.FINISHED;
-            }
+            return super.processOutput(out);
         }
     }
 
-    boolean needsPostProcessing() {
-        return this.tempOut != null && this.tempADTNode != null;
-    }
-
-    ADTNode<ADTState<I, O>, I, O> getCurrentADTNode() {
-        return this.currentADTNode;
-    }
-
-    O getTempOut() {
-        return this.tempOut;
+    @Override
+    protected void resetProgress() {
+        this.asIndex = 0;
     }
 
     ADTTransition<I, O> getTransition() {
